+++ ui.py
@@ -21,29 +21,25 @@ class GPREFINE_PT_stroke_refine_panel(GPR_refine, Panel):
     def draw(self, context):
         layout = self.layout
 
-class GPREFINE_PT_targets(GPR_refine, Panel):
-    bl_label = "Targets"
+class GPREFINE_PT_target_filter(GPR_refine, Panel):
+    bl_label = "Target Filter"
     bl_parent_id = "GPREFINE_PT_stroke_refine_panel"
     bl_options = {'DEFAULT_CLOSED'}
 
-    # def draw_header(self, context):
-    #     layout = self.layout
-    #     layout.prop(context.scene.gprsettings, 'use_select', text='')
+    def draw_header(self, context):
+        layout = self.layout
+        layout.prop(context.scene.gprsettings, 'use_custom_targets', text='')
 
     def draw(self, context):
         layout = self.layout
         layout.use_property_split = True # send properties to the right side
-
-        col_pref = layout.column()
-        col_pref.prop(context.scene.gprsettings, 'use_context')
-        col_pref.prop(context.scene.gprsettings, 'use_select')
         
         col_filter = layout.column()
-        col_filter.prop(context.scene.gprsettings, 'layer_tgt')
-        col_filter.prop(context.scene.gprsettings, 'frame_tgt')
-        col_filter.prop(context.scene.gprsettings, 'stroke_tgt')
+        col_filter.prop(context.scene.gprsettings, 'layer_tgt', text="Layer")
+        col_filter.prop(context.scene.gprsettings, 'frame_tgt', text="Frame")
+        col_filter.prop(context.scene.gprsettings, 'stroke_tgt', text="Stroke")
         
-        col_filter.active = not (context.scene.gprsettings.use_select or (context.scene.gprsettings.use_context and context.mode == 'PAINT_GREASE_PENCIL'))
+        col_filter.active = context.scene.gprsettings.use_custom_targets
 
 
 class GPREFINE_PT_Selector(GPR_refine, Panel):
@@ -375,7 +371,7 @@ class GPREFINE_PT_auto_join(GPR_refine, Panel):
 
 classes = (
 GPREFINE_PT_stroke_refine_panel,
-GPREFINE_PT_targets,
+GPREFINE_PT_target_filter,
 GPREFINE_PT_Selector,
 GPREFINE_PT_stroke_shape_refine,